very very last fix

+++ tasks/seq/varfolomeev_g_quick_sort_simple_merge/func_tests/main.cpp
@@ -35,7 +35,7 @@ TEST(varfolomeev_g_quick_sort_simple_merge_seq, Test_getRandomVector_func) {
   int lower_bound = -50;
   int upper_bound = 50;
   std::vector<int> vec = varfolomeev_g_quick_sort_simple_merge_seq::getRandomVector_seq(size, lower_bound, upper_bound);
-  EXPECT_EQ(vec.size(), size);
+  EXPECT_EQ((int)vec.size(), size);
   for (int value : vec) {
     EXPECT_GE(value, lower_bound);
     EXPECT_LE(value, upper_bound);
@@ -49,7 +49,7 @@ TEST(varfolomeev_g_quick_sort_simple_merge_seq, Test_getAntisortedVecor_func) {
   int size = 100;
   int start = -50;
   std::vector<int> vec = varfolomeev_g_quick_sort_simple_merge_seq::getAntisorted_seq(size, start);
-  EXPECT_EQ(vec.size(), size);
+  EXPECT_EQ((int)vec.size(), size);
   for (int i = 0; i < (int)vec.size() - 1; i++) {
     EXPECT_LE(vec[i + 1], vec[i]);
   }
